requirements.txt
Django==4.1.3
dj-database-url==2.1.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
gunicorn==23.0.0
whitenoise==6.7.0
Pillow==10.4.0
